feat: header refactor and page identification logic #29
