+++ src/components/Burger/Burger.jsx
@@ -1,6 +1,9 @@
 import React from "react";
+import usePageIdentification from "../../hooks/usePageIdentification";
 
 const Burger = () => {
+  const { isMainPage } = usePageIdentification();
+
   return <div>burger</div>;
 };
