+++ src/components/Header/Header.css
@@ -11,86 +11,6 @@
   background: transparent;
 }
 
-.header__links {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 30px;
-}
-
-.header__links_movies {
-  gap: 20px;
-}
-
-.header__link {
-  text-decoration: none;
-  padding: 0;
-  margin: 0;
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 16px;
-  color: #fff;
-}
-
-.header__link_movies {
-  font-size: 14px;
-  line-height: 18px;
-  color: #000;
-}
-
-.header__link_movies-white {
-  color: #fff;
-}
-
-.header__link:visited {
-  text-decoration: none;
-}
-
-.header__button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 76px;
-  height: 32px;
-  border-radius: 3px;
-  background-color: #2BE080;
-  color: #000;
-}
-
-.header__account-button {
-  margin: 0 0 0 30px;
-  padding: 0 3px;
-  display: flex;
-  flex-flow: row nowrap;
-  align-items: center;
-  justify-content: flex-end;
-  gap: 7px;
-  color: #000;
-  height: 32px;
-  width: 100px;
-  background: #FFFFFF;
-  box-shadow: 0 3px 14px rgba(0, 0, 0, 0.08);
-  border-radius: 40px;
-  text-decoration: none;
-}
-
-.header__account-button-text {
-  padding: 0;
-  margin: 0;
-  font-weight: 500;
-  font-size: 13px;
-  line-height: 12px;
-}
-
-.header__account-icon {
-  background: url(./../../images/ava.svg) no-repeat center center;
-  height: 26px;
-  width: 26px;
-}
-
-.header__link:hover, .header__account-button:hover, .header__link:active, .header__button:hover, .header__button:active {
-  opacity: 0.5;
-}
 
 @media (max-width: 768px) {
   .header {
@@ -101,15 +21,5 @@
 @media (max-width: 420px) {
   .header {
     padding: 0 14px;
-  }
-  .header__links {
-    gap: 14px;
-  }
-  .header__link{
-    font-size: 10px;
-  }
-  .header__button {
-    width: 54px;
-    height: 26px;
   }
 }
